--- tests/neo.UnitTests/Persistence/UT_CloneCache.cs
@@ -1,37 +1,0 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Neo.SmartContract;
-
-namespace Neo.UnitTests.Persistence
-{
-    [TestClass]
-    public class UT_CloneCache
-    {
-        [TestMethod]
-        public void TestCacheOverrideIssue2572()
-        {
-            var snapshot = TestBlockchain.GetTestSnapshot();
-            var storages = snapshot.CreateSnapshot();
-
-            storages.Add
-                (
-                new StorageKey() { Key = new byte[] { 0x00, 0x01 }, Id = 0 },
-                new StorageItem() { Value = new byte[] { } }
-                );
-            storages.Add
-                (
-                new StorageKey() { Key = new byte[] { 0x01, 0x01 }, Id = 0 },
-                new StorageItem() { Value = new byte[] { 0x05 } }
-                );
-
-            storages.Commit();
-
-            var item = storages.GetAndChange(new StorageKey() { Key = new byte[] { 0x01, 0x01 }, Id = 0 });
-            item.Value = new byte[] { 0x06 };
-
-            var res = snapshot.TryGet(new StorageKey() { Key = new byte[] { 0x01, 0x01 }, Id = 0 });
-            Assert.AreEqual("05", res.Value.ToHexString());
-            res = storages.TryGet(new StorageKey() { Key = new byte[] { 0x01, 0x01 }, Id = 0 });
-            Assert.AreEqual("06", res.Value.ToHexString());
-        }
-    }
-}
